Add passing environment_config to stack template render
It would be useful to have environment_config available in the stack
config template for access to parameters like region. It would be also
a good place to define extra attributes that could be used for default
values across environment.

Adds passing enviornment_config as "environment_config" to stack
Config read.